backend/pyproject.toml
[tool.poetry]
name = "pyproject"
version = "0.1.0"
description = ""
authors = ["juanMaAV92 <juanmanuel.armero@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0.0"
matplotlib = "^3.7.1"
obspy = "^1.4.0"
fpdf2 = "^2.7.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
